+++ native/Sources/CaptureObjC/ExceptionCatch.m
@@ -1,0 +1,10 @@
+#import "ExceptionCatch.h"
+
+NSException *_Nullable GDCatchException(void(NS_NOESCAPE ^block)(void)) {
+    @try {
+        block();
+        return nil;
+    } @catch (NSException *exception) {
+        return exception;
+    }
+}